Refactor: Remove terms2 module and update references
- Deleted the terms2 module which contained various term definitions.
- Updated all references in the grammar, language, and test files to use the terms module instead of terms2.
- Ensured that all term-related functionality remains intact after the refactor.
